Copters: Added MultiWii Pro EZ3 Black board (From ReadyToFlyQuads.com)
Copters: GPS_AUTO: 4800 replaced by 115200

+++ libraries/AP_GPS/GPS.h
@@ -210,7 +210,7 @@ protected:
     bool _have_raw_velocity;
 
 	// detected baudrate
-	uint16_t _baudrate;
+	uint32_t _baudrate;
 
 private:
